+++ packages/react-components/src/ModalProvider/ModalManager.tsx
@@ -1,7 +1,0 @@
-import { useModalState } from "./hooks/useModalState";
-
-export function ModalManager(): JSX.Element {
-    const modals = useModalState();
-
-    return <div id="modal-root">{modals}</div>;
-}