+++ packages/react-components/src/ModalProvider/ModalManager/ModalManager.styles.ts
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const ModalManagerRoot = styled.div(({ theme }) => ({
+    position: "fixed",
+    zIndex: theme.zIndex.modal + 1,
+}));