+++ packages/react-components/src/ModalProvider/ModalManager/ModalManager.tsx
@@ -1,0 +1,8 @@
+import { useModalState } from "../hooks/useModalState";
+import { ModalManagerRoot } from "./ModalManager.styles";
+
+export function ModalManager(): JSX.Element {
+    const modals = useModalState();
+
+    return <ModalManagerRoot id="modal-root">{modals}</ModalManagerRoot>;
+}