+++ packages/react-components/src/Pagination/PaginationItem/PaginationItem.tsx
@@ -27,6 +27,7 @@ const PaginationItem = ({
             )}
             isEllipsis={isEllipsis}
             as={isEllipsis ? "div" : "button"}
+            type="button"
             {...rest}
         />
     );